refactor: split rolling dependency selection
